server: future-proof realisation caching
previously, we would just create a directory for each target
which has both a valid input hash, output hashs and dumps in
the store. but the amount of subdirectories per directory is
often extremely small for our use-case, so here we insert a
quite critical optimization, which, instead of a directory,
creates just a symlink for all targets which only produce
the default output, which should be the majority.
Thus, we avoid reaching the limit of subdirs really early.
(we expect ~100'000 targets for a normal full-system build)